rental_service/requirements.txt
djangorestframework~=3.14.0
psycopg2~=2.9.3
